+++ src/LakeWright.Conversations/ConversationsServiceCollectionExtensions.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using LakeWright.Core.Features;
 
 namespace LakeWright.Conversations;
 
@@ -32,6 +33,7 @@ public static class ConversationsServiceCollectionExtensions
             new GenieSharedSpaceOptionsValidator(provider.GetService<ILoggerFactory>()));
 
         services.AddSingleton(TimeProvider.System);
+        services.TryAddSingleton<ILakeWrightFeatureGate, AlwaysOnFeatureGate>();
         services.TryAddSingleton<IConversationOwnership, MemoryConversationOwnership>();
 
         services.AddHttpClient<IGenieConversations, GenieConversations>((provider, client) =>